Add flickr30k_captions

+++ lavis/models/blip2_models/blip2.py
@@ -239,7 +239,7 @@ def compute_sim_matrix(model, data_loader, **kwargs):
     image_embeds = []
     for samples in data_loader:
         image = samples["image"]
-        image = image.to(model.device).half()
+        image = image.to(model.device)
         image_feat, vit_feat = model.forward_image(image)
         image_embed = model.vision_proj(image_feat)
         image_embed = F.normalize(image_embed, dim=-1)
